Adds clock, buttons transparency. Fixes paths to images

+++ KeyInterceptor/KeyReleasedEventArgs.cs
@@ -7,9 +7,9 @@ namespace KeyInterceptor
 	{
 		public Keys KeyCode { get; }
 		public DateTime PressedTimestamp { get; }
-		public long Duration { get; }
+		public double Duration { get; }
 
-		public KeyReleasedEventArgs(Keys keyCode, DateTime pressedTimestamp, long milliseconds)
+		public KeyReleasedEventArgs(Keys keyCode, DateTime pressedTimestamp, double milliseconds)
 		{
 			KeyCode = keyCode;
 			PressedTimestamp = pressedTimestamp;
